pyproject.toml
[project]
name = "font-renderer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "deepthought", email = "deepthough@postech.ac.kr" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "pillow>=11.2.1",
]

[project.scripts]
font-renderer = "font_renderer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
